Moved testing library into new namespace, refactored some code to be more shared in test-library-agnostic classes/traits
